Refactored Modals as components
